64-bit support
add support for 64-bit

+++ ManagedSpyLib/Mem.h
@@ -18,7 +18,7 @@ namespace Microsoft {
 namespace ManagedSpy {
 
 	struct SharedData {
-		SIZE_T Size;
+		unsigned int Size;
 		BYTE Data[1];
 	};
